moved over our value type to float instead of double to make it more compact

+++ src/main/java/com/vishnuvyas/classifiers/Classifier.java
@@ -8,7 +8,7 @@ import java.util.Map;
  * Created by vishnu on 4/20/14.
  */
 public interface Classifier<VectorType, LabelType> {
-    public Map<LabelType,Double> predictions(VectorType point);
+    public Map<LabelType,Float> predictions(VectorType point);
 
     public void save(OutputStream os);
     public void load(InputStream is);
